timetable/requirements.txt
aiohttp==3.9.5
alembic==1.13.3
fastapi==0.115.0
pydantic==2.9.2
PyJWT==2.3.0
SQLAlchemy==2.0.35
uvicorn==0.31.0
asyncio==3.4.3
python-dotenv==1.0.1
pydantic_settings==2.6.1
bcrypt==4.2.0
python-multipart==0.0.17
cffi==1.17.1 
cryptography==43.0.3 
pycparser==2.22
asyncpg==0.30.0
aio-pika==9.5.4